requirements.txt
amqp==2.2.2
asn1crypto==0.23.0
bcrypt==3.1.4
billiard==3.5.0.3
celery==4.1.0
cffi==1.11.2
cryptography==2.1.4
Django==2.0
django-celery-beat==1.1.0
django-configurations==2.0
ephem==3.7.6.0
Fabric3==1.13.1.post1
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
idna==2.6
kombu==4.1.0
olefile==0.44
paramiko==2.4.0
Pillow==4.3.0
psycopg2==2.7.3.2
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.5.3
PyNaCl==1.2.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
vine==1.1.4
